backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
websockets==12.0
openai==1.3.5
requests==2.31.0
python-multipart==0.0.6
aiohttp==3.9.1
pydantic==2.5.2
python-jose==3.3.0 